dockerfiles/dist_training/requirements.txt
torch
torchvision
lightning
tensorboardX
pandas
pillow
timm
h5py
scikit-learn
